% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{average_pages_read}
\alias{average_pages_read}
\title{Average Pages Calculation}
\usage{
average_pages_read(data)
}
\arguments{
\item{data}{Dataset of filtered books.}
}
\value{
Average number of pages per day over the remaining year.
}
\description{
Calculates the average pages read per day if a person reads the top listed books from the dataset.
    The number of books is based on a calculation of 1 book a week for the remaining year.
}
\examples{
average_pages_read(data.frame(NumberOfPages = c(25, 168, 150, 210)))
}
