% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{filter_data}
\alias{filter_data}
\title{Filter Book Data}
\usage{
filter_data(min_rating, max_pages, min_publish_date, max_publish_date)
}
\arguments{
\item{min_rating}{Minimum average rating.}

\item{max_pages}{Maximum number of pages.}

\item{min_publish_date}{Minimum publication date.}

\item{max_publish_date}{Maximum publication date.}
}
\value{
Filtered set of data.
}
\description{
Filters book data for a minimum rating, maximum number of pages,
    and publication date.
}
\examples{
data <- filter_data(4, 200, as.Date("2020-01-01"), as.Date("2020-09-30"))
}
